Move event log steps into appropriate module Remove unnecessary context use in case api look up steps
