+++ acceptance_tests/features/steps/case_look_up.py
@@ -4,56 +4,55 @@ from random import randint
 from uuid import uuid4
 
 import requests
-from behave import then, given, step
+from behave import then, given
 from structlog import wrap_logger
 
-from acceptance_tests.utilities.test_case_helper import tc
 from config import Config
 
 logger = wrap_logger(logging.getLogger(__name__))
-caseapi_url = f'{Config.CASEAPI_SERVICE}/cases/'
+case_api_url = f'{Config.CASEAPI_SERVICE}/cases/'
 
 
 @then("a case can be retrieved from the case API service")
 def get_case_by_id(context):
     case_id = context.case_created_events[0]['payload']['collectionCase']['id']
 
-    response = requests.get(f'{caseapi_url}{case_id}')
+    response = requests.get(f'{case_api_url}{case_id}')
 
     assert response.status_code == 200, 'Case not found'
 
 
 @given('a random caseId is generated')
 def generate_random_uuid(context):
-    context.dummy_case_id = uuid4()
-    context.test_endpoint_with_non_existent_value = context.dummy_case_id
-    logger.info(f'Dummy caseId = {context.dummy_case_id}')
+    dummy_case_id = uuid4()
+    context.test_endpoint_with_non_existent_value = dummy_case_id
+    logger.info(f'Dummy caseId = {dummy_case_id}')
 
 
 @given('a random uprn is generated')
 def generate_random_uprn(context):
-    context.random_uprn = randint(15000000000, 19999999999)
-    context.test_endpoint_with_non_existent_value = f'uprn/{context.random_uprn}'
-    logger.info(f'Dummy uprn = {context.random_uprn}')
+    random_uprn = randint(15000000000, 19999999999)
+    context.test_endpoint_with_non_existent_value = f'uprn/{random_uprn}'
+    logger.info(f'Dummy uprn = {random_uprn}')
 
 
 @given('a random caseRef is generated')
-def generate_random_caseref(context):
-    context.random_caseref = randint(15000000, 19999999)
-    context.test_endpoint_with_non_existent_value = f'ref/{context.random_caseref}'
-    logger.info(f'Dummy caseRef = {context.random_caseref}')
+def generate_random_case_ref(context):
+    random_case_ref = randint(15000000, 19999999)
+    context.test_endpoint_with_non_existent_value = f'ref/{random_case_ref}'
+    logger.info(f'Dummy caseRef = {random_case_ref}')
 
 
 @then('case API should return a 404 when queried')
 def get_non_existent_case_id(context):
-    response = requests.get(f'{caseapi_url}{context.test_endpoint_with_non_existent_value}')
+    response = requests.get(f'{case_api_url}{context.test_endpoint_with_non_existent_value}')
 
     assert response.status_code == 404, 'Case returned'
 
 
 @then('case API returns multiple cases for a UPRN')
 def find_multiple_cases_by_uprn(context):
-    response = requests.get(f'{caseapi_url}uprn/10008677190')
+    response = requests.get(f'{case_api_url}uprn/10008677190')
     response.raise_for_status()
 
     response_data = json.loads(response.content)
@@ -67,62 +66,15 @@ def find_multiple_cases_by_uprn(context):
 
 
 @then('a case can be retrieved by its caseRef')
-def find_case_by_caseRef(context):
+def find_case_by_case_ref(context):
     case_ref = context.case_created_events[0]['payload']['collectionCase']['caseRef']
 
-    response = requests.get(f'{caseapi_url}ref/{case_ref}')
+    response = requests.get(f'{case_api_url}ref/{case_ref}')
 
     assert response.status_code == 200, 'Case ref not found'
 
 
-@then('events of type "{event_type}" are logged against the case')
-def correct_event_type_logged(context, event_type):
-    for case in context.case_created_events:
-        case_id = case['payload']['collectionCase']['id']
-        response = requests.get(f'{caseapi_url}{case_id}?caseEvents=true').content.decode("utf-8")
-        response_json = json.loads(response)
-
-        for caseEvent in response_json['caseEvents']:
-            if caseEvent['eventType'] == event_type:
-                return
-        assert False
-
-
-@then("the events logged for the case are {event_type_list}")
-def event_logged_for_case(context, event_type_list):
-    from acceptance_tests.features.steps.event_log import check_if_event_list_is_exact_match
-    check_if_event_list_is_exact_match(event_type_list, context.emitted_case['id'])
-
-
-def _check_if_event_is_logged(expected_event_type, actual_logged_events):
-    actual_logged_event_types = [event['eventType'] for event in actual_logged_events]
-    tc.assertEqual(actual_logged_event_types.count(expected_event_type), 1,
-                   msg=(f'Expected event type = {expected_event_type},'
-                        f' actual logged event types = {actual_logged_event_types}'))
-
-
-@step('"{event_type}" events are logged against the cases included in the reminder')
-def check_print_case_selected_event_is_logged_against_reminder_cases(context, event_type):
-    for case_id in context.reminder_case_ids:
-        actual_logged_events = get_logged_events_for_case_by_id(case_id)
-        _check_if_event_is_logged(event_type, actual_logged_events)
-
-
-@then("events logged for receipted cases are {event_type_list}")
-def event_logged_for_receipting(context, event_type_list):
-    actual_logged_events = get_logged_events_for_case_by_id(context.emitted_case_updated['id'])
-    _check_if_event_list_is_exact_match(event_type_list, actual_logged_events)
-
-
 def get_logged_events_for_case_by_id(case_id):
-    response = requests.get(f'{caseapi_url}{case_id}?caseEvents=true').content.decode("utf-8")
+    response = requests.get(f'{case_api_url}{case_id}?caseEvents=true').content.decode("utf-8")
     response_json = json.loads(response)
     return response_json['caseEvents']
-
-
-def _check_if_event_list_is_exact_match(event_type_list, actual_logged_events):
-    expected_logged_event_types = event_type_list.replace('[', '').replace(']', '').split(',')
-    actual_logged_event_types = [event['eventType'] for event in actual_logged_events]
-
-    tc.assertCountEqual(expected_logged_event_types, actual_logged_event_types,
-                        msg="Actual logged event types did not match expected")
