+++ acceptance_tests/features/steps/event_log.py
@@ -2,6 +2,7 @@ from behave import step
 from retrying import retry
 
 from acceptance_tests.features.steps.case_look_up import get_logged_events_for_case_by_id
+from acceptance_tests.utilities.test_case_helper import tc
 
 
 @step("events logged against the case are {event_type_list}")
@@ -10,19 +11,31 @@ def correct_event_types_logged(context, event_type_list):
         check_if_event_list_is_exact_match(event_type_list, case['payload']['collectionCase']['id'])
 
 
+@step('"{event_type}" events are logged against the cases included in the reminder')
+def check_print_case_selected_event_is_logged_against_reminder_cases(context, event_type):
+    for case_id in context.reminder_case_ids:
+        actual_logged_events = get_logged_events_for_case_by_id(case_id)
+        _check_if_event_is_logged(event_type, actual_logged_events)
+
+
+@step("events logged for receipted cases are {event_type_list}")
+def event_logged_for_receipting(context, event_type_list):
+    actual_logged_events = get_logged_events_for_case_by_id(context.emitted_case_updated['id'])
+    check_if_event_list_is_exact_match(event_type_list, actual_logged_events)
+
+
 @retry(stop_max_attempt_number=10, wait_fixed=1000)
 def check_if_event_list_is_exact_match(event_type_list, case_id):
     actual_logged_events = get_logged_events_for_case_by_id(case_id)
     expected_logged_event_types = event_type_list.replace('[', '').replace(']', '').split(',')
-    expected_logged_event_types_copy = expected_logged_event_types.copy()
+    actual_logged_event_types = [event['eventType'] for event in actual_logged_events]
 
-    assert len(actual_logged_events) == len(expected_logged_event_types), "wrong number of events logged"
+    tc.assertCountEqual(expected_logged_event_types, actual_logged_event_types,
+                        msg="Actual logged event types did not match expected")
 
-    for case_event in actual_logged_events:
-        for expected_event in expected_logged_event_types:
-            if case_event['eventType'] == expected_event:
-                expected_logged_event_types_copy.remove(expected_event)
-                break
 
-    if len(expected_logged_event_types_copy) > 0:
-        assert False, "didn't log expected event types"
+def _check_if_event_is_logged(expected_event_type, actual_logged_events):
+    actual_logged_event_types = [event['eventType'] for event in actual_logged_events]
+    tc.assertEqual(actual_logged_event_types.count(expected_event_type), 1,
+                   msg=(f'Expected event type = {expected_event_type},'
+                        f' actual logged event types = {actual_logged_event_types}'))
